requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
joblib
requests
google-generativeai>=0.7.0
python-dotenv>=0.19.0
